YTDB-635: extract IndexOrderedPlanner from MatchExecutionPlanner
Shrinks MatchExecutionPlanner.class from ~125KB back to ~113KB by moving
index-ordered detection, multi-source mode selection, and the supporting
helpers (resolveOrderByToAliasProperty, isFilteredScanLikelyWorthwhile,
hasSingleRowGuarantee, etc.) into a new package-private IndexOrderedPlanner.

Profiling IC2/IC10 showed MatchExecutionPlanner growth triggered a JIT
de-inlining cascade: HashMap.getNode inclusive samples went from 1651 to
3125 (+89%), HashMap.hash from 92 to 528 (+474%). Re-sizing the planner
class restores the inlining budget for shared utilities on the hot
compile path — the index-ordered code itself is cold-path (runs once per
query at planning time) and only loses one virtual dispatch.
